🎨style: PrivacySheet 디자인 추가
